+++ frontend/src/main.tsx
@@ -4,6 +4,7 @@ import App from "./App";
 import UpdateCenter from "./UpdateCenter";
 import { I18nProvider } from "./i18n";
 import "./style.css";
+import "./motion.css";
 createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <I18nProvider>